requirements.txt
pandas>=2.2.0
requests>=2.32.0
python-dotenv>=1.0.1
pyarrow>=16.0.0
